+++ EQEmuServer/zone/questmgr.cpp
@@ -1575,7 +1575,7 @@ void QuestManager::clearspawntimers() {
 			char errbuf[MYSQL_ERRMSG_SIZE];
 			char *query = 0;
 			database.RunQuery(query, MakeAnyLenString(&query, "DELETE FROM respawn_times WHERE id=%lu AND "
-				"instance_id=%lu",iterator.GetData()->GetID(), zone->GetInstanceID()), errbuf);
+				"instance_id=%lu",(unsigned long)iterator.GetData()->GetID(), (unsigned long)zone->GetInstanceID()), errbuf);
 			safe_delete_array(query);		
 			iterator.Advance();
 		}
